# Remove inner components from DCEL faces