## Description

This greatly simplifies the overlay algorithm, in particular with the
"reassign face" step.

It requires a pre-processing stage where ghost edges are created to make
each input geometry fully connected. It also adds a ghost line between
the two geometries.

## Check List

Have you:

- Added unit tests? Yes

- Add cmprefimpl tests? (if appropriate?) Yes, uses existing tests.

## Related Issue

- https://github.com/peterstace/simplefeatures/issues/284

## Benchmark Results

Benchmarks are marginally faster (note that benchmark coverage isn't great so far, and a performance improvement wasn't the intended or expected here).

```
name                    old time/op    new time/op    delta
Intersection/n=10-4        141µs ± 8%     142µs ± 9%    ~     (p=0.780 n=15+14)
Intersection/n=100-4      1.20ms ± 7%    1.10ms ± 4%  -8.15%  (p=0.000 n=15+14)
Intersection/n=1000-4     13.4ms ± 9%    12.2ms ±10%  -9.22%  (p=0.000 n=14+15)
Intersection/n=10000-4     153ms ± 6%     141ms ±11%  -7.52%  (p=0.000 n=14+15)

name                    old alloc/op   new alloc/op   delta
Intersection/n=10-4       56.2kB ± 0%    58.9kB ± 0%  +4.79%  (p=0.000 n=15+15)
Intersection/n=100-4       492kB ± 0%     473kB ± 0%  -3.76%  (p=0.000 n=15+15)
Intersection/n=1000-4     6.05MB ± 0%    5.86MB ± 0%  -3.24%  (p=0.000 n=15+15)
Intersection/n=10000-4    58.9MB ± 0%    56.1MB ± 0%  -4.65%  (p=0.000 n=15+13)

name                    old allocs/op  new allocs/op  delta
Intersection/n=10-4          713 ± 0%       749 ± 0%  +5.09%  (p=0.000 n=15+15)
Intersection/n=100-4       5.22k ± 0%     5.18k ± 0%  -0.83%  (p=0.000 n=15+11)
Intersection/n=1000-4      49.7k ± 0%     49.4k ± 0%  -0.59%  (p=0.000 n=15+15)
Intersection/n=10000-4      494k ± 0%      492k ± 0%  -0.41%  (p=0.000 n=15+15)
```